+++ verticapy/_config/colors.py
@@ -20,7 +20,40 @@ from typing import Union, Optional
 import matplotlib.colors as plt
 
 import verticapy._config.config as conf
-from verticapy._config.config import COLORS_OPTIONS
+
+COLORS_OPTIONS: dict[str, list] = {
+    "rgb": ["red", "green", "blue", "orange", "yellow", "gray"],
+    "sunset": ["#36688D", "#F3CD05", "#F49F05", "#F18904", "#BDA589"],
+    "retro": ["#A7414A", "#282726", "#6A8A82", "#A37C27", "#563838"],
+    "shimbg": ["#0444BF", "#0584F2", "#0AAFF1", "#EDF259", "#A79674"],
+    "swamp": ["#6465A5", "#6975A6", "#F3E96B", "#F28A30", "#F05837"],
+    "med": ["#ABA6BF", "#595775", "#583E2E", "#F1E0D6", "#BF9887"],
+    "orchid": ["#192E5B", "#1D65A6", "#72A2C0", "#00743F", "#F2A104"],
+    "magenta": ["#DAA2DA", "#DBB4DA", "#DE8CF0", "#BED905", "#93A806"],
+    "orange": ["#A3586D", "#5C4A72", "#F3B05A", "#F4874B", "#F46A4E"],
+    "vintage": ["#80ADD7", "#0ABDA0", "#EBF2EA", "#D4DCA9", "#BF9D7A"],
+    "vivid": ["#C0334D", "#D6618F", "#F3D4A0", "#F1931B", "#8F715B"],
+    "berries": ["#BB1924", "#EE6C81", "#F092A5", "#777CA8", "#AFBADC"],
+    "refreshing": ["#003D73", "#0878A4", "#1ECFD6", "#EDD170", "#C05640"],
+    "summer": ["#728CA3", "#73C0F4", "#E6EFF3", "#F3E4C6", "#8F4F06"],
+    "tropical": ["#7B8937", "#6B7436", "#F4D9C1", "#D72F01", "#F09E8C"],
+    "india": ["#F1445B", "#65734B", "#94A453", "#D9C3B1", "#F03625"],
+    "default": ["#FE5016", "#263133", "#0073E7", "#FDE159", "#33C180", "#FF454F"],
+}
+
+
+def color_validator(val: Union[str, list, None]) -> Literal[True]:
+    if (
+        (isinstance(val, str) and val in COLORS_OPTIONS)
+        or isinstance(val, list)
+        or val == None
+    ):
+        return True
+    else:
+        raise ValueError(
+            "The option must be a list of colors, None, or in"
+            f" [{'|'.join(COLORS_OPTIONS)}]"
+        )
 
 
 def get_colors(d: Optional[dict] = {}, idx: Optional[int] = None) -> Union[list, str]:
@@ -47,7 +80,9 @@ def get_colors(d: Optional[dict] = {}, idx: Optional[int] = None) -> Union[list,
         return colors[idx % len(colors)]
 
 
-def get_cmap(color: str = "", reverse: bool = False) -> plt.LinearSegmentedColormap:
+def get_cmap(
+    color: Optional[str] = None, reverse: bool = False
+) -> plt.LinearSegmentedColormap:
     if not (color):
         cm1 = plt.LinearSegmentedColormap.from_list(
             "verticapy_cmap", ["#FFFFFF", get_colors()[0]], N=1000
@@ -69,3 +104,7 @@ def get_cmap(color: str = "", reverse: bool = False) -> plt.LinearSegmentedColor
             return plt.LinearSegmentedColormap.from_list(
                 "verticapy_cmap", ["#FFFFFF", color], N=1000
             )
+
+
+colors_option = conf.Option("colors", None, "", color_validator, COLORS_OPTIONS)
+conf.register_option(colors_option)
