+++ verticapy/_utils/_sql/_collect.py
@@ -55,10 +55,9 @@ Returns
 bool
     True if the operation succeeded, False otherwise.
     """
-    if not (conf.get_option["save_query_profile"]) or (
-        isinstance(conf.get_option["save_query_profile"], list)
-        and name not in conf.get_option["save_query_profile"]
-    ):
+    value = conf.get_option("save_query_profile")
+
+    if not (value):
         return False
     try:
